Organize some enum types
