+++ app/src/main/java/com/example/scorekeeper/game/SortingOrder.kt
@@ -1,0 +1,9 @@
+package com.example.scorekeeper.game
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class SortingOrder(val readableName: String) {
+    ALPHABETICAL(readableName = "Alphabetical Order"),
+    REVERSE_ALPHABETICAL(readableName = "Reverse Alphabetical Order")
+}
