+++ app/src/main/java/com/example/scorekeeper/game/types/Game.kt
@@ -1,6 +1,7 @@
 package com.example.scorekeeper.game.types
 
 import com.example.scorekeeper.game.Player
+import com.example.scorekeeper.game.SortingOrder
 import com.example.scorekeeper.game.renderers.GameRenderer
 import kotlinx.serialization.Polymorphic
 import kotlinx.serialization.Serializable
@@ -13,10 +14,4 @@ sealed class Game {
     abstract var isComplete: Boolean
 
     abstract fun getRenderer(): GameRenderer
-}
-
-@Serializable
-enum class SortingOrder(val readableName: String) {
-    ALPHABETICAL(readableName = "Alphabetical Order"),
-    REVERSE_ALPHABETICAL(readableName = "Reverse Alphabetical Order")
 }
