+++ app/src/main/java/com/example/scorekeeper/game/types/PointGame.kt
@@ -1,6 +1,7 @@
 package com.example.scorekeeper.game.types
 
 import com.example.scorekeeper.game.Player
+import com.example.scorekeeper.game.SortingOrder
 import com.example.scorekeeper.game.renderers.GameRenderer
 import com.example.scorekeeper.game.renderers.PointGameRenderer
 import kotlinx.serialization.SerialName
