+++ app/src/main/java/com/example/scorekeeper/game/types/RankedRoundGame.kt
@@ -1,6 +1,7 @@
 package com.example.scorekeeper.game.types
 
 import com.example.scorekeeper.game.Player
+import com.example.scorekeeper.game.SortingOrder
 import com.example.scorekeeper.game.round.Round
 import com.example.scorekeeper.game.renderers.GameRenderer
 import com.example.scorekeeper.game.renderers.RankedRoundGameRenderer
